service-unit:refactor: move task handler to application
